[olarft] Eliminate dangling service/helper files outside of the /utils folder
